fix type conversion to pass Windows build

+++ testing/utils/TestUtilsCWriter.c
@@ -85,7 +85,7 @@ int main(int argc, char *argv[])
     adios2_define_attribute(ioH, "name", adios2_type_string, "TestUtilsCWrite",
                             1);
     adios2_define_attribute(ioH, "nwriters", adios2_type_int, &nproc, 1);
-    unsigned short shape2D[2] = {d2_Nx, d2_Ny};
+    unsigned short shape2D[2] = {(unsigned short)d2_Nx, (unsigned short)d2_Ny};
     adios2_define_attribute(ioH, "shape2D", adios2_type_unsigned_short, shape2D,
                             2);
